refatorar: substitua nodemailer pela função sendEmail em vários módulos correção: atualize o assunto do email para verificação e emails de alerta correção: altere o método de análise de float para int para valor de depósito

+++ Controller/Controller-Deposito.ts
@@ -24,7 +24,7 @@ export default class Deposito {
             res.status(400).json({ message: "Valores inválidos" });
             return;
         }
-        valor = parseFloat(valor);
+        valor = parseInt(valor);
         idconta = parseInt(idconta);
         iddeposito = parseInt(iddeposito);
